Don't run mod-kb-ebsco tests before deploying to production
Whenever this build is run, we run our rubocop style checker, and then
run the entire mod-kb-ebsco suite

https://travis-ci.org/thefrontside/folio-mod-kb-ebsco-deploy/builds/330868207

But this work isn't necessary however because the test suite is being
run as part of the actual repo build. In fact, because master is
protected by the test suite on the mod-kb-ebsco repository, it's
impossible for any changes to actually be seen by this deployment repo
unless they've already run (and passed) the test suite.

This removes those checks as entirely redundant so that we can have
faster merge-to-deployment times.