Harden mod-kb-ebsco deployment pipeline
Herein lies work dedicated to increasing the reliability of our CD
pipeline for mod-kb-ebsco.  Since deployments to our infrastructure
are decoupled from the build & deploy process of
folio-org/mod-kb-ebsco, there is a long and winding road from the
merge button to our kubernetes cluster.

Bringing the reliability of the pipeline to 100% will require
changes outside the reaches of this repository, but for now this
seeks to serve as a bucket for any improvements we can make in this
neighborhood.

This initial commit augments the deploy script to force a rolling
update on the mod-kb-ebsco deployment whenever a build is fired on
Circle.  This will grab the 'latest' tag from DockerHub, which may not
always be correct.